pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ciphercourt"
version = "0.1.0"
description = "A data availability audit framework for pre-match tennis modeling"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "CipherCourt Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ciphercourt = "ciphercourt.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ciphercourt*"]
exclude = ["tests*"]
